fixed font issue
Because of react weirdness, it seems nothing actually inherits from body. When the font-family is declared at the level of the div with id root it get passed down to the children elements effectively.
It is strange that it was working intermittently before, but now I can reload this page all day and the font sticks around